feat: add status for job application and move config to cmd
